Function Call and search fucntionality update
calls to working_title() replaced with calls to date_specified_numbers(). 

Casts to int re-added to generate map function.

When searching for data from the last n weeks/months the data is now added to the output as an additional column and when passed into generate_map() will display the change within that timeframe rather than the total cases/deaths of counties that met the condition. Additionally, when passed to export_data() this information will appear in the resulting file.